+++ src/domain-model/service-registration.js
@@ -1,5 +1,6 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { BoundedContextRegistry } from './bounded-contexts/bounded-context-registry.js';
+import { CoreEntityRegistry } from './entities/core-entity-registry.js';
 import { DomainOverviewDescriptor } from './overview/domain-overview-descriptor.js';
 
 export function addDomainModel(services) {
@@ -8,6 +9,7 @@ export function addDomainModel(services) {
   }
 
   services.registerSingleton('BoundedContextRegistry', () => new BoundedContextRegistry());
+  services.registerSingleton('CoreEntityRegistry', () => new CoreEntityRegistry());
   services.registerSingleton('DomainOverviewDescriptor', () => new DomainOverviewDescriptor());
 
   return services;